requirements.txt
python-dotenv
boto3
mysql-connector-python
pydantic
requests
pymongo